// packages/components/sections/features/src/types.ts
export interface MainImage {
  image: string;
  alt?: string;
}
// ...
export interface FeaturedItem {
  description?: string;
  mainImage?: MainImage;
  title?: string;
  subtitle?: string;
  _key?: string;
  _type?: string;
}